Refactor editor capabilities
